Seperate movement from framerate

+++ engine/Core.h
@@ -25,11 +25,11 @@ namespace gfx {
         // GETTERS & SETTERS
         int width();
         int height();
-        Color px(int x, int y);
-        void px(math::Matrix<double> m, Color color, bool motionBlur=false);
+        Color px(float x, float y);
+        void px(math::Matrix<float> m, Color color, bool motionBlur=false);
         void px(Point p, Color color);
-        void px(int x, int y, Color color, bool motionBlur=false);
-        void px(int x, int y, Uint8 r, Uint8 g, Uint8 b, bool motionBlur=false);
+        void px(float x, float y, Color color, bool motionBlur=false);
+        void px(float x, float y, Uint8 r, Uint8 g, Uint8 b, bool motionBlur=false);
         void factor(float factor);
 
     private:
